pyproject.toml
[project]
name = "kinpdfle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "img2pdf>=0.6.1",
    "pillow>=11.2.1",
]
